fix: enforce canonical provider construction
